requirements.txt
bs4
jinja2
lxml
openai
python-dotenv
weaviate-client